Closes #4.
First refactor commit. Builds route map
